+++ packages/std/src/string/pad.ts
@@ -1,3 +1,5 @@
+import validateStringCount from './validateStringCount';
+
 /**
  * Pads string on the left and right sides if it's shorter than length.
  * Padding characters are truncated if they can't be evenly divided by length.
@@ -11,6 +13,8 @@
  * pad('abc', 3) //=> 'abc'
  */
 export default function pad(str: string, length = 0, chars = ' '): string {
+  validateStringCount(length, 'Length');
+
   const strLength = str.length;
   if (strLength >= length) {
     return str;